Refactor of the importing from postman collection
Refactoring make the importing less error-prone.
I wanted to achieve the situation where really simple and
on the first side uncompleted collection
would be parsed anyway in the state it is.
I saw also that collections in many situations are throwing when some data
aren't filled. Additional thing was importer utilities - it wasn't really
helpful and it didn't solve the multi-line comments in a good way. I
decided to change it to contains prefix for all the lines.